server/botc/requirements.txt
aiogram==2.25.1
aiosqlite==0.17.0
python-dotenv==1.0.0
pytz==2023.3